refactor(module-resolver): expose public interfaces
